# specs/009-investment-workbench/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Lucking 投资工作台 API
  version: 1.0.0
  description: 前端唯一公共契约；所有端点均为同域 /api/v1 路径。
servers:
  - url: /api/v1
security:
  - sessionCookie: []
paths:
...
  /stocks:
    get:
      summary: 查询股票列表
      operationId: listStocks
      parameters:
        - name: query
          in: query
          schema: { type: string, maxLength: 80 }
        - name: venue_code
          in: query
          schema: { type: string, enum: [XSHG, XSHE, XBSE] }
        - name: listing_status
          in: query
          schema: { type: string }
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: 稳定分页的股票列表
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StockListResponse' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /stocks/{stock_id}:
    get:
      summary: 获取股票详情和最新行情状态
      operationId: getStock
      parameters:
        - $ref: '#/components/parameters/StockId'
      responses:
        '200':
          description: 股票详情
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StockDetailResponse' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /stocks/{stock_id}/daily-quotes:
    get:
      summary: 获取股票日线行情区间
      operationId: listDailyQuotes
      parameters:
        - $ref: '#/components/parameters/StockId'
        - name: start_date
          in: query
          schema: { type: string, format: date }
        - name: end_date
          in: query
          schema: { type: string, format: date }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 400, default: 120 }
      responses:
        '200':
          description: 按交易日升序的行情；停牌日不合成记录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DailyQuoteListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /j-gold/research:
    get:
      summary: 获取 J金股研究驾驶舱
      operationId: getJGoldResearch
      parameters:
        - name: recommendation_month
          in: query
          schema: { type: string, format: date }
        - name: broker_name
          in: query
          schema: { type: string, maxLength: 160 }
        - name: industry
          in: query
          schema: { type: string, maxLength: 160 }
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: 同一筛选范围的驾驶舱结果
          content:
            application/json:
              schema: { $ref: '#/components/schemas/JGoldResearchResponse' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /j-gold/stocks/{stock_id}:
    get:
      summary: 获取可追溯的 J金股股票详情
      operationId: getJGoldStockResearch
      parameters:
        - $ref: '#/components/parameters/StockId'
        - name: recommendation_month
          in: query
          required: true
          schema: { type: string, format: date }
      responses:
        '200':
          description: 股票研究详情
          content:
            application/json:
              schema: { $ref: '#/components/schemas/JGoldStockResearchResponse' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: lucking_session
  parameters:
...
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 1000, default: 100 }
    Offset:
      name: offset
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
...
    StockId:
      name: stock_id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    BadRequest:
      description: 请求字段或业务范围无效
      content:
        application/json:
          schema: { $ref: '#/components/schemas/BadRequestErrorResponse' }
    Unauthorized:
      description: 未认证或会话已失效
      content:
        application/json:
          schema: { $ref: '#/components/schemas/UnauthorizedErrorResponse' }
...
    NotFound:
      description: 资源不存在或不属于当前用户
      content:
        application/json:
          schema: { $ref: '#/components/schemas/NotFoundErrorResponse' }
...
    InternalServerError:
      description: 未预期服务错误；不得返回堆栈、SQL、秘密或第三方原始响应
      content:
        application/json:
          schema: { $ref: '#/components/schemas/InternalServerErrorResponse' }
  schemas:
    BusinessErrorCode:
      type: integer
      description: 稳定业务错误码；HTTP 状态表达协议结果，业务码表达应用原因
      enum:
        - 100001
        - 100002
        - 100003
        - 200001
        - 200002
        - 200003
        - 300001
        - 400001
        - 400002
        - 400003
        - 400004
        - 500001
        - 500002
        - 900001
      x-enum-varnames:
        - INVALID_CREDENTIALS
        - SESSION_INVALID
        - CSRF_VALIDATION_FAILED
        - REQUEST_VALIDATION_FAILED
        - QUERY_RANGE_INVALID
        - PASSWORD_POLICY_VIOLATION
        - RESOURCE_NOT_FOUND
        - IMPORTANT_DATE_CONFLICT
        - WATCHLIST_NAME_CONFLICT
        - WATCHLIST_MEMBER_CONFLICT
        - WATCHLIST_CAPACITY_EXCEEDED
        - RATE_LIMITED
        - EXTERNAL_DEPENDENCY_UNAVAILABLE
        - INTERNAL_SERVER_ERROR
    ErrorDetail:
      type: object
      required: [field, code, message]
      properties:
        field: { type: [string, 'null'] }
        code: { type: string }
        message: { type: string }
    Pagination:
      type: object
      required: [limit, offset, total, has_more]
      properties:
        limit: { type: integer, minimum: 1 }
        offset: { type: integer, minimum: 0 }
        total: { type: integer, minimum: 0 }
        has_more: { type: boolean }
    SuccessResponseBase:
      type: object
      required: [code, message, errors, request_id, timestamp]
      properties:
        code: { type: integer, const: 0 }
        message: { type: string, const: '' }
        errors:
          type: array
          maxItems: 0
          items: { $ref: '#/components/schemas/ErrorDetail' }
        request_id: { type: string, minLength: 1 }
        timestamp: { type: string, format: date-time }
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [code, message, data, errors, request_id, timestamp]
      properties:
        code: { $ref: '#/components/schemas/BusinessErrorCode' }
        message: { type: string, minLength: 1 }
        data: { type: 'null' }
        errors:
          type: array
          items: { $ref: '#/components/schemas/ErrorDetail' }
        request_id: { type: string, minLength: 1 }
        timestamp: { type: string, format: date-time }
    BadRequestErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            code: { type: integer, enum: [200001, 200002, 200003] }
    UnauthorizedErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            code: { type: integer, enum: [100001, 100002] }
...
    NotFoundErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            code: { type: integer, const: 300001 }
...
    InternalServerErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            code: { type: integer, const: 900001 }
...
    StockSummary:
      type: object
      required: [stock_id, market_code, venue_code, security_code, name, listing_status]
      properties:
        stock_id: { type: string, format: uuid }
        market_code: { type: string }
        venue_code: { type: string }
        security_code: { type: string }
        name: { type: string }
        listing_status: { type: string }
    StockDetail:
      allOf:
        - $ref: '#/components/schemas/StockSummary'
        - type: object
          required: [market_data_status]
          properties:
            latest_quote: { oneOf: [{ $ref: '#/components/schemas/DailyQuote' }, { type: 'null' }] }
            market_data_status: { type: string, enum: [CURRENT, STALE, MISSING] }
    DailyQuote:
      type: object
      required: [trade_date, open, high, low, close, pre_close, change, pct_chg, vol, amount, updated_at]
      properties:
        trade_date: { type: string, format: date }
        open: { type: string }
        high: { type: string }
        low: { type: string }
        close: { type: string }
        pre_close: { type: string }
        change: { type: string }
        pct_chg: { type: string }
        vol: { type: string }
        amount: { type: string }
        updated_at: { type: string, format: date-time }
...
    StockPageData:
      type: object
      required: [items, pagination]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/StockSummary' }
        pagination: { $ref: '#/components/schemas/Pagination' }
...
    StockListResponse:
      unevaluatedProperties: false
      allOf:
        - $ref: '#/components/schemas/SuccessResponseBase'
        - type: object
          required: [data]
          properties:
            data: { $ref: '#/components/schemas/StockPageData' }
    StockDetailResponse:
      unevaluatedProperties: false
      allOf:
        - $ref: '#/components/schemas/SuccessResponseBase'
        - type: object
          required: [data]
          properties:
            data: { $ref: '#/components/schemas/StockDetail' }
    DailyQuoteListResponse:
      unevaluatedProperties: false
      allOf:
        - $ref: '#/components/schemas/SuccessResponseBase'
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/DailyQuote' }
...
    JGoldResearchData:
      type: object
      required: [metrics, items, pagination, signals, industries, broker_ability, diffusion, quality, selected_month, available_months]
      properties:
        metrics: { type: object, additionalProperties: true }
        items:
          type: array
          items: { type: object, additionalProperties: true }
        pagination: { $ref: '#/components/schemas/Pagination' }
        signals: { type: array, items: { type: object, additionalProperties: true } }
        industries: { type: array, items: { type: object, additionalProperties: true } }
        broker_ability: { type: array, items: { type: object, additionalProperties: true } }
        diffusion: { type: array, items: { type: object, additionalProperties: true } }
        quality: { type: object, additionalProperties: true }
        selected_month: { type: string, format: date }
        available_months: { type: array, items: { type: string, format: date } }
    JGoldStockResearchData:
      type: object
      required: [stock, recommendations, history, price_basis, source, generated_at, quality]
      properties:
        stock: { $ref: '#/components/schemas/StockSummary' }
        industry: { type: [string, 'null'] }
        recommendations: { type: array, items: { type: object, additionalProperties: true } }
        history: { type: array, items: { type: object, additionalProperties: true } }
        latest_quote_date: { type: [string, 'null'], format: date }
        price_basis: { type: string }
        source: { type: string }
        generated_at: { type: string, format: date-time }
        quality: { type: string }
    JGoldResearchResponse:
      unevaluatedProperties: false
      allOf:
        - $ref: '#/components/schemas/SuccessResponseBase'
        - type: object
          required: [data]
          properties:
            data: { $ref: '#/components/schemas/JGoldResearchData' }
    JGoldStockResearchResponse:
      unevaluatedProperties: false
      allOf:
        - $ref: '#/components/schemas/SuccessResponseBase'
        - type: object
          required: [data]
          properties:
            data: { $ref: '#/components/schemas/JGoldStockResearchData' }